#Refactor feat(api): Implement caching and fix initial tab selection
Refactors the `useFetch` composable to accept custom TanStack/Vue Query options, making it more flexible.

Utilizes this new functionality to set a 5-minute `staleTime` on fetches for tabbed and latest posts. This improves performance by caching post data and reducing unnecessary API requests.

Fixes a bug in the `CountryTabs` component where the active tab would not be set on navigating back to the page. The new caching prevented the watcher from firing. The watcher is now set to `{ immediate: true }` to ensure it runs on component mount, correctly setting the tab from cached data.