Add error state for latest posts
Adds an error handling state to the "Latest Posts" section on the home page.
Creates a new `LatestPostsError.vue` component to display a user-friendly message when posts fail to load.
The `useGetLatestPosts` composable's `isError` flag is now used to conditionally render this new component.
This improves robustness by providing clear feedback to the user in case of an API error.